refactor(matrix): split monitor handler pipeline
